fix: move IRecipeKnowledge and fix deserialization encapsulation
- Moved IRecipeKnowledge from api.capability to api package (no longer
  part of capability system)
- Removed getRecipes() method that exposed internal mutable set
- Removed addRecipe() method that bypassed validation logic
- Added deserializeInto() method for efficient in-place deserialization
- Updated Villager to use deserializeInto() instead of wasteful copy loop

This prevents potential data corruption from exposed mutable state and
eliminates unnecessary object allocation during deserialization.
